issue 3.2 solved and all bugs and conflicts fixed
