--- COMPLETION_REPORT.md
@@ -1,14 +1,0 @@
-US-0.1 Completion Report
-=========================
-
-Summary:
-- Implemented an idempotent CLI to validate environment, verify or run migrations, and ensure an admin user exists.
-
-What was verified:
-- `npm run db:migrate` completes after migration fix.
-- `npm run init` creates/updates admin user and reports temporary password when needed.
-- `npm run db:seed` updated to hash passwords and seeds dev data.
-
-Outstanding recommendations:
-- Replace AES-CBC encrypted env handling with AEAD (e.g., AES-GCM) or add HMAC.
-- Review `package-lock.json` consistency before pushing.
