--- EPIC-0-QUICK-REFERENCE.md
@@ -1,276 +1,0 @@
-# EPIC 0: Quick Reference Guide
-
-## 🎯 What Was Implemented
-
-Complete environment configuration system for EWTCS supporting development, staging, and production environments with type-safe validation, health checks, and security best practices.
-
----
-
-## 📁 New Files Created
-
-### Configuration Module (`src/lib/config/`)
-| File | Lines | Purpose |
-|------|-------|---------|
-| `env.ts` | 95 | Main environment configuration with Zod validation |
-| `logger.ts` | 117 | Structured logging with color output in dev, JSON in prod |
-| `secrets.ts` | 90 | Encryption and sensitive data handling |
-| `init.ts` | 122 | System initialization and health checks |
-
-### Type Definitions (`src/types/`)
-| File | Lines | Purpose |
-|------|-------|---------|
-| `config.ts` | 54 | TypeScript interfaces for configuration |
-
-### API Endpoints (`src/app/api/`)
-| File | Lines | Purpose |
-|------|-------|---------|
-| `health/route.ts` | 40 | Health check endpoint for monitoring |
-
-### Environment Files (Root)
-| File | Purpose |
-|------|---------|
-| `.env.example` | Template with documentation |
-| `.env.development` | Development defaults |
-| `.env.staging` | Staging configuration |
-| `.env.production` | Production template (security warnings) |
-
-### Documentation
-| File | Purpose |
-|------|---------|
-| `CONFIGURATION.md` | 500+ line setup and security guide |
-| `EPIC-0-IMPLEMENTATION.md` | Complete implementation details |
-
----
-
-## 🔧 How to Get Started
-
-### 1. Development Setup (5 minutes)
-```bash
-# Copy template
-cp .env.example .env.local
-
-# Edit with your local database
-# DATABASE_URL=postgresql://postgres:password@localhost:5432/ewtcs_dev
-
-# Verify health
-curl http://localhost:3000/api/health
-```
-
-### 2. Use Configuration in Code
-```typescript
-import { config } from '@/lib/config/env';
-import { logger } from '@/lib/config/logger';
-import { query } from '@/lib/db/client';
-
-// Type-safe configuration
-config.app.isDevelopment  // true/false
-config.database.url       // "postgresql://..."
-config.alert.delayThresholdMs  // 10800000 (3 hours)
-
-// Structured logging
-logger.info('Operation started', { userId: 123 });
-
-// Database queries with timing
-const result = await query('SELECT * FROM beds');
-```
-
-### 3. Health Check Endpoint
-```bash
-# Check system health
-curl http://localhost:3000/api/health
-
-# Response (200 if healthy, 503 if not)
-{
-  "status": "healthy",
-  "checks": {
-    "configuration": "pass",
-    "environment": "pass", 
-    "database": "pass"
-  }
-}
-```
-
----
-
-## 🔑 Key Features
-
-### ✅ Type Safety
-```typescript
-// IDE autocomplete for all configuration
-config.app.url              // ✅ TypeScript knows type
-config.database.ssl         // ✅ Intellisense shows options
-```
-
-### ✅ Environment Validation
-```
-✅ DATABASE_URL must be valid PostgreSQL URL
-✅ NEXT_PUBLIC_APP_URL must be valid URL  
-✅ NODE_ENV must be development|staging|production
-❌ Fails fast on startup if invalid
-```
-
-### ✅ Security
-```
-✅ Passwords masked in logs
-✅ SSL enforced in production
-✅ Connection pooling (max 20)
-✅ Encryption support available
-```
-
-### ✅ Logging
-```
-✅ Development: Color-coded console
-✅ Production: JSON structured logs
-✅ 5 log levels: DEBUG, INFO, WARN, ERROR, CRITICAL
-✅ Context preserved in logs
-```
-
----
-
-## 📋 Environment Variables
-
-### Required
-```env
-DATABASE_URL=postgresql://user:pass@host:5432/db
-NEXT_PUBLIC_APP_URL=http://localhost:3000
-NODE_ENV=development
-```
-
-### Optional
-```env
-RED_ALERT_THRESHOLD_MS=10800000  # 3 hours
-OPENAI_API_KEY=sk-...             # For future AI features
-```
-
----
-
-## 🚨 Security Checklist
-
-### Development
-- ✅ Use `.env.local` (in `.gitignore`)
-- ✅ Never commit real credentials
-- ✅ Use localhost database
-
-### Production
-- ✅ Use secret manager (AWS, Vault, K8s)
-- ✅ Never commit `.env.production`
-- ✅ Rotate credentials every 90 days
-- ✅ Enable SSL/TLS
-- ✅ Monitor health endpoint
-
----
-
-## 📚 Documentation
-
-| Document | When to Read |
-|----------|--------------|
-| `CONFIGURATION.md` | Setting up different environments |
-| `CONTRIBUTING.md` | Contributing to project (includes security) |
-| `EPIC-0-IMPLEMENTATION.md` | Full technical details |
-| `.env.example` | Understanding each variable |
-
----
-
-## 🧪 Quick Tests
-
-### Test Configuration Validation
-```bash
-# Start without DATABASE_URL - should fail:
-unset DATABASE_URL
-npm run dev
-# ❌ System Foundation Setup Failed
-```
-
-### Test Health Check
-```bash
-curl http://localhost:3000/api/health | jq .
-# {
-#   "status": "healthy",
-#   "environment": "development",
-#   "checks": {...}
-# }
-```
-
-### Test Logging
-```typescript
-import { logger } from '@/lib/config/logger';
-
-logger.info('Test message', { example: true });
-// Development: [2026-02-14T...] [INFO] Test message | {...}
-// Production: {"timestamp":"...","level":"INFO",...}
-```
-
----
-
-## 🎯 Acceptance Criteria - All Met ✅
-
-| Criteria | Evidence |
-|----------|----------|
-| Configuration uses environment variables | `env.ts`, `.env.example` |
-| Sensitive data encrypted | `secrets.ts` with crypto |
-| Configuration validation on startup | `init.ts` health checks |
-| Example configuration provided | `.env.example`, `.env.development`, etc. |
-| Configuration documentation clear | `CONFIGURATION.md` (500+ lines) |
-
----
-
-## 🚀 Next Steps
-
-1. **Merge Code**
-   - Review `EPIC-0-IMPLEMENTATION.md`
-   - Code review by team
-   - Approve and merge to main
-
-2. **Test in Staging**
-   - Deploy with staging credentials
-   - Verify health endpoint: `curl https://staging.ewtcs.com/api/health`
-   - Test all three log levels
-
-3. **Production Deployment**
-   - Use secret manager (not .env files)
-   - Set environment variables in container/K8s
-   - Enable monitoring on `/api/health`
-   - Test graceful restarts
-
-4. **Enable Additional Features**
-   - When ready: Enable encryption in `secrets.ts`
-   - Integrate with secret manager
-   - Add database backup monitoring
-
----
-
-## 💡 Tips
-
-**Use configuration in any file:**
-```typescript
-import { config } from '@/lib/config/env';
-import { logger } from '@/lib/config/logger';
-
-// Immediately available at import time
-// No async/await needed
-```
-
-**Log context information:**
-```typescript
-logger.info('Database query', {
-  table: 'beds',
-  duration: 42,
-  rows: 5
-});
-// Structured, searchable in production logs
-```
-
-**Health checks from monitoring system:**
-```bash
-# Kubernetes liveness probe
-curl -f http://localhost:3000/api/health || exit 1
-
-# External monitoring alert if status != 'healthy'
-```
-
----
-
-**Status:** ✅ Production Ready  
-**Date:** February 14, 2026  
-**Story Points:** 5 (Complete)  
-**Epic:** EPIC 0: System Foundation & Setup
