--- reports/PR-DESCRIPTION.md
@@ -1,110 +1,0 @@
-## feat: Emergency Bed Grid Dashboard (US-1.1) #6
-
-### 📋 Summary
-
-Implements real-time bed status dashboard for emergency ward nurses. View all beds at a glance with color-coded stages, automatic delay detection, and filter capabilities.
-
-**User Story:** As a nurse, I want to see all emergency beds in a grid layout so I can quickly understand bed status without switching systems.
-
----
-
-## ✨ Key Features
-
-- **📊 Responsive Grid Layout** - Auto-adjusts 1-4 columns based on screen size
-- **🎨 Color-Coded Stages** - 8 visual indicators (Empty → Triage → Registration → Doctor → Treatment → Decision → Discharge → Cleaning)
-- **⏱️ Real-Time Tracking** - Automatic elapsed time calculation since patient admission
-- **🚨 Red Alert System** - Visual warnings for beds exceeding 3-hour threshold
-- **🔍 Filter Toggle** - Show delayed beds only
-- **🔄 Manual Refresh** - Update bed statuses on demand
-
----
-
-## 🏗️ Implementation
-
-**Architecture:** Feature-first structure (`src/features/bed-dashboard/`)
-
-**Performance Optimizations:**
-- React.memo on components (20-30% fewer re-renders)
-- useMemo for expensive calculations
-- useCallback for event handlers
-- Optimized SQL queries with indexes
-
-**Database:**
-- Migration 005: Creates `beds`, `stages`, `bed_stage_logs` tables
-- 8 pre-seeded workflow stages
-- Indexed columns for fast queries
-
----
-
-## ✅ Acceptance Criteria (5/5)
-
-- ✅ All beds visible without scrolling
-- ✅ Bed numbers clearly displayed
-- ✅ Responsive grid layout
-- ✅ Current stage names shown
-- ✅ Real-time updates (refresh on demand)
-
----
-
-## 🧪 Testing
-
-```bash
-✅ Build: Successful (14.6s, 0 errors)
-✅ Lint: 0 errors, 0 warnings
-✅ TypeScript: Strict mode, no errors
-✅ File sizes: All under 200 lines
-✅ Manual testing: All features working
-```
-
----
-
-## 📁 Files Changed
-
-**Created (11 files):**
-- Components: BedCard, BedGrid, BedStatusLegend, BedDashboardClient
-- Logic: bed-actions.ts, bed-queries.ts, stage-queries.ts, utils.ts
-- Types: bed.types.ts
-- Database: migration 005
-
-**Modified (4 files):**
-- dashboard/page.tsx - Integrated dashboard
-- audit.ts - Fixed schema bug
-- not-found.tsx - Created 404 page
-- useDatabaseStatus.ts - Fixed lint warning
-
-**Documentation:**
-- US-1.1-IMPLEMENTATION-REPORT.md (75+ sections)
-- README.md updated
-
----
-
-## 🚀 How to Test
-
-```bash
-npm run db:migrate  # Create tables
-npm run db:seed     # Add sample data
-npm run dev         # Start server
-# Login: admin / Admin@123
-# Visit: http://localhost:3000/dashboard
-```
-
----
-
-## ✅ Checklist
-
-- [x] Code follows style guidelines
-- [x] All files under 200 lines
-- [x] Build passes (14.6s)
-- [x] Lint passes (0 errors)
-- [x] Documentation complete
-- [x] All acceptance criteria met
-- [x] Ready for production
-
----
-
-**Closes #6**
-
-**Related:** EPIC 1 - Nurse Desk Bed Dashboard  
-**Story Points:** 5 | **Priority:** P0 (Critical)
-
-**Full technical details:** See [US-1.1-IMPLEMENTATION-REPORT.md](US-1.1-IMPLEMENTATION-REPORT.md)
