--- reports/README.md
@@ -1,63 +1,0 @@
-# Implementation Reports
-
-This folder contains detailed implementation reports and documentation for completed user stories and issues.
-
-## Purpose
-
-- Keep implementation documentation separate from the main codebase
-- Provide detailed records of how features were implemented
-- Store PR descriptions and technical details for reference
-- Maintain a history of development work
-
-## Contents
-
-Each report typically includes:
-- User story/issue details
-- Acceptance criteria verification
-- Implementation approach
-- Files changed
-- Testing instructions
-- Security and compliance notes
-
-## Naming Convention
-
-Reports should follow the format: `US-X.Y-IMPLEMENTATION-REPORT.md` or `ISSUE-XX-REPORT.md`
-
----
-
-## 📄 Available Reports
-
-### EPIC 1: Nurse Desk Bed Dashboard
-- **[US-1.1-IMPLEMENTATION-REPORT.md](US-1.1-IMPLEMENTATION-REPORT.md)** - View All Emergency Beds in Grid Layout
-  - Status: ✅ Complete (Feb 16, 2026)
-  - Responsive grid layout with color-coded beds
-  - Elapsed time tracking and delay detection
-  - Filter functionality and statistics dashboard
-
-### EPIC 3: Time Tracking & Stage Logging
-- **[US-3.1-IMPLEMENTATION-REPORT.md](US-3.1-IMPLEMENTATION-REPORT.md)** - Capture Patient Entry Time
-  - Status: ✅ Complete (Feb 17, 2026)
-  - Server-side admission time capture
-  - Immutable entry timestamps using database time
-
-### EPIC 5: Authentication & Role-Based Access
-- **[US-5.3-IMPLEMENTATION-REPORT.md](US-5.3-IMPLEMENTATION-REPORT.md)** - User Management (Admin)
-  - Status: ✅ Complete (Feb 15, 2026)
-  - Create, edit, activate/deactivate users
-  - Admin dashboard with audit logging
-  - Full CRUD operations with validation
-
-### Architecture & Planning
-- **[FEATURE-FIRST-ARCHITECTURE-PLAN.md](FEATURE-FIRST-ARCHITECTURE-PLAN.md)** - Feature-First Hybrid Architecture
-  - Comprehensive architecture documentation
-  - Directory structure guidelines
-  - Best practices and patterns
-  - Migration from monolithic to feature-based
-
-- **[PR-DESCRIPTION.md](PR-DESCRIPTION.md)** - Pull Request Templates
-  - Standard PR description format
-  - Checklist for reviewers
-
----
-
-**Note:** These reports are for documentation purposes and can be referenced in PRs but are not part of the production codebase.
