--- reports/US-3.1-IMPLEMENTATION-REPORT.md
@@ -1,80 +1,0 @@
-# US-3.1: Capture Patient Entry Time - Implementation Report
-
-**Issue**: US-3.1 - Capture Patient Entry Time  
-**Epic**: EPIC 3 - Time Tracking & Stage Logging  
-**Status**: Completed  
-**Story Points**: 3  
-**Priority**: P0 (Critical)  
-**Completion Date**: February 17, 2026
-
----
-
-## Overview
-
-Implemented server-side capture of patient admission time when a bed transitions from a non-patient stage (Empty/Cleaning) to an occupied stage. The timestamp is generated by the database server, stored as a timezone-aware value, and remains immutable until the bed is cleared back to a non-patient stage.
-
----
-
-## Acceptance Criteria - All Met
-
-| Criterion | Status | Implementation |
-|-----------|--------|----------------|
-| Timestamp captured when bed changes from Empty to Occupied | Met | Set in SQL during stage update when `patient_start_time` is NULL |
-| Timestamp uses server time (not client time) | Met | `NOW()` used in PostgreSQL update |
-| Timestamp stored in ISO 8601 format | Met | Stored in `TIMESTAMPTZ` and returned as ISO 8601 |
-| Timestamp immutable once set | Met | SQL keeps existing value once set |
-| Timezone recorded with timestamp | Met | `TIMESTAMPTZ` with server timezone |
-
----
-
-## Implementation Details
-
-### 1. Server-Side Mutation Logic
-
-**File**: `src/features/bed-dashboard/lib/bed-mutations.ts`
-
-- Uses `NOW()` in SQL to set admission time when a bed becomes occupied.
-- Keeps `patient_start_time` unchanged if already set.
-- Clears `patient_start_time` only when bed returns to non-patient stages (Empty/Cleaning).
-- Returns the stored timestamp from the database after update.
-
-**Core SQL logic:**
-```
-patient_start_time = CASE
-  WHEN $2 THEN NULL
-  WHEN patient_start_time IS NULL THEN NOW()
-  ELSE patient_start_time
-END
-```
-
----
-
-## Files Changed
-
-- `src/features/bed-dashboard/lib/bed-mutations.ts`
-
----
-
-## Testing Performed
-
-Manual verification steps:
-1. Move a bed from Empty -> Occupied stage. Confirm `patient_start_time` is set.
-2. Move the same bed through other occupied stages. Confirm `patient_start_time` does not change.
-3. Move bed to Cleaning or Empty. Confirm `patient_start_time` is cleared.
-
----
-
-## Security and Compliance
-
-- Timestamp is generated on the server via Postgres `NOW()` (no client input).
-- Audit trail continues to log stage transitions through existing bed stage logs.
-
----
-
-## Deployment Notes
-
-No schema changes required. This is a logic update only.
-
----
-
-**Implementation Status**: Complete and ready for review
